Refactored read DTC functions to have less repetitive code by using a helper function.
Could still improve more by not having separate calls for each sub-function, and instead build the request in the read_dtc_info function and call the helper from there. Everything about calling the sub-function is uniform outside of building the request and trimming to response to not include useless info.